+++ src/index.ts
@@ -9,6 +9,7 @@ import StIcon from './components/icon/index.vue';
 import StInput from './components/input/index.vue';
 import { StNotificationService } from './components/notification/service';
 import StNotificationsGroup from './components/notifications-group/index.vue';
+import StPagination from './components/pagination/index.vue';
 import StPopover from './components/popover/index.vue';
 import StRow from './components/row/index.vue';
 import StTextarea from './components/textarea/index.vue';
@@ -25,6 +26,7 @@ const install = (Vue: VueConstructor) => {
   Vue.component('st-popover', StPopover);
   Vue.component('st-row', StRow);
   Vue.component('st-textarea', StTextarea);
+  Vue.component('st-pagination', StPagination);
 
   /* eslint-disable-next-line no-param-reassign */
   Vue.prototype.$stNotification = StNotificationService;
@@ -43,4 +45,5 @@ export default {
   StPopover,
   StRow,
   StTextarea,
+  StPagination,
 };