+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 # Changelog
 
-## 0.2.0
+## 0.2.1
 
 ### Features
 * Added `Radio` component to `Offers` theme
 * Added new component: `Switch`
 * Added new component: `Notifications`
 * Added new component: `Scrollbar`
+* Added new component: `Pagination`
 * Added new component: `Popover`
 * Added new component: `Badge`
 
@@ -14,6 +15,7 @@
 * Fixed `Offers` theme for `Button` component
 * Fixed `Checkbox` value type 
 * Added components generator. If you are about to make a new component just run `npm run generator:component` and follow the instructions.
+* Prepared for publishing